Add REST API mapping, service revamp
-Fix encoding and decoding functionality
-Add rest api mapping with swagger 2 support
-Add global exception handler and own exceptions
-Change the outcome of find all request - now it doesn't show the
password, it's obtainable only by direct request (by website or uuid)

+++ src/main/java/com/github/sebastiancegielka/service/impl/PasswordEntryServiceImpl.java
@@ -2,8 +2,10 @@ package com.github.sebastiancegielka.service.impl;
 
 import com.github.sebastiancegielka.dto.PasswordEntryCreateDTO;
 import com.github.sebastiancegielka.dto.PasswordEntryDTO;
+import com.github.sebastiancegielka.dto.PasswordEntryFindAllDTO;
+import com.github.sebastiancegielka.exception.EntryNotFoundException;
 import com.github.sebastiancegielka.mapper.PasswordEntryMapper;
-import com.github.sebastiancegielka.model.SecretKey;
+import com.github.sebastiancegielka.model.EncodingKey;
 import com.github.sebastiancegielka.model.PasswordEntry;
 import com.github.sebastiancegielka.repository.PasswordEntryRepository;
 import com.github.sebastiancegielka.service.PasswordEntryService;
@@ -11,7 +13,8 @@ import com.github.sebastiancegielka.utility.PasswordUtility;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
+import javax.crypto.SecretKey;
+import java.security.NoSuchAlgorithmException;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
@@ -32,17 +35,11 @@ public class PasswordEntryServiceImpl implements PasswordEntryService {
     }
 
     @Override
-    public PasswordEntryDTO findOneById(Long id) {
-        PasswordEntry entry = repository.findById(id).orElseThrow(RuntimeException::new);
-        return mapper.toEntryDTO(passwordUtility.decryptPassword(entry));
-    }
-
-    @Override
     public PasswordEntryDTO findOneByUuid(String uuid) {
         PasswordEntry entry = repository.findByUuid(uuid);
         if (entry != null) {
             return mapper.toEntryDTO(passwordUtility.decryptPassword(entry));
-        } else throw new RuntimeException();
+        } else throw new EntryNotFoundException("uuid", uuid);
     }
 
     @Override
@@ -51,7 +48,7 @@ public class PasswordEntryServiceImpl implements PasswordEntryService {
         Set<PasswordEntry> decryptedEntry = new HashSet<>();
 
         if(entry.isEmpty()){
-            throw new RuntimeException();
+            throw new EntryNotFoundException("website", website);
         }
         for (PasswordEntry passwordEntry : entry) {
             decryptedEntry.add(passwordUtility.decryptPassword(passwordEntry));
@@ -62,46 +59,48 @@ public class PasswordEntryServiceImpl implements PasswordEntryService {
     }
 
     @Override
-    public Set<PasswordEntryDTO> findAll() {
+    public Set<PasswordEntryFindAllDTO> findAll() {
         List<PasswordEntry> entryList = repository.findAll();
-        List<PasswordEntry> decryptedEntryList = new ArrayList<>();
-        for (PasswordEntry passwordEntry : entryList) {
-            decryptedEntryList.add(passwordUtility.decryptPassword(passwordEntry));
-        }
-        return decryptedEntryList.stream()
-                .map(x->mapper.toEntryDTO(x))
+
+        return entryList.stream()
+                .map(x->mapper.toEntryFindAllDTO(x))
                 .collect(Collectors.toSet());
     }
 
     @Override
-    public PasswordEntryDTO createEntry(PasswordEntryCreateDTO entry) {
+    public PasswordEntryFindAllDTO createEntry(PasswordEntryCreateDTO entry) {
         PasswordEntry newEntry = new PasswordEntry();
         if(passwordUtility.validate(entry)){
-            PasswordEntryCreateDTO checkedEntry = passwordUtility.checkPassword(entry);
+            try {
+                PasswordEntryCreateDTO checkedEntry = passwordUtility.checkPassword(entry);
 
-            String password = checkedEntry.getPassword();
-            String key = passwordUtility.generateString(8);
+                String password = checkedEntry.getPassword();
+                SecretKey key = passwordUtility.generateKey();
+                String encryptedPass = passwordUtility.encryptPassword(password, key);
 
-            String encryptedPass = passwordUtility.encryptPassword(password, key);
+                newEntry = mapper.toEntry(entry);
 
-            newEntry = mapper.toEntry(entry);
+                EncodingKey encryptKey = new EncodingKey();
+                encryptKey.setSecretKey(key);
+                encryptKey.setEntry(newEntry);
 
-            SecretKey encryptKey = new SecretKey();
-            encryptKey.setSecretKey(key);
-            encryptKey.setEntry(newEntry);
+                newEntry.setPassword(encryptedPass);
+                newEntry.setEncodingKey(encryptKey);
 
-            newEntry.setPassword(encryptedPass);
-            newEntry.setSecretKey(encryptKey);
-
-            repository.save(newEntry);
+                repository.save(newEntry);
+            } catch (NoSuchAlgorithmException e) {
+                e.printStackTrace();
+            }
         }
-        return mapper.toEntryDTO(newEntry);
+        return mapper.toEntryFindAllDTO(newEntry);
     }
 
     @Override
-    public void deleteEntryById(Long id) {
-        repository.deleteById(id);
+    public void deleteEntryByUuid(String uuid) {
+        PasswordEntry entry = repository.findByUuid(uuid);
+        if(entry != null) {
+            Long id = entry.getId();
+            repository.deleteById(id);
+        } else throw new EntryNotFoundException("uuid", uuid);
     }
-
-
 }
